feat: 기본 애플리케이션 구조 및 컴포넌트 추가
- index.html, jsconfig.json, package.json, vite.config.js 파일 추가로 기본 애플리케이션 구조 설정
- src 디렉토리에 app.js, main.js, setupTests.js 파일 추가하여 초기화 및 테스트 환경 구성
- CartDisplay, CartEventHandler, CartItem, HelpModal, NotificationBar, OrderSummary, ProductSelector, StockInfo, DiscountEngine, PointsCalculator, PriceCalculator, StockCalculator 등 다양한 컴포넌트 및 헬퍼 모듈 구현
- 각 컴포넌트에 대한 단위 테스트 추가로 코드 안정성 및 품질 향상

✅ 새로운 애플리케이션 구조로 유지보수성 및 확장성 향상
