+++ apps/basic/src/components/__tests__/CartItem.test.js
