+++ apps/basic/src/components/__tests__/ProductSelector.test.js
