+++ apps/basic/src/components/__tests__/StockInfo.test.js
