+++ apps/basic/src/constants/UIConstants.js
