+++ apps/basic/src/helpers/PriceCalculator.js
