+++ apps/basic/src/helpers/PromotionManager.js
