+++ apps/basic/src/helpers/__tests__/PointsCalculator.test.js
