+++ apps/basic/src/helpers/__tests__/StockCalculator.test.js
