+++ apps/basic/vite.config.js
@@ -1,0 +1,21 @@
+import { resolve } from 'path';
+import { defineConfig } from 'vitest/config';
+
+export default defineConfig({
+  resolve: {
+    alias: {
+      '@': resolve(__dirname, 'src'),
+      '@components': resolve(__dirname, 'src/components'),
+      '@helpers': resolve(__dirname, 'src/helpers'),
+      '@constants': resolve(__dirname, 'src/constants'),
+      '@core': resolve(__dirname, 'src/core'),
+      '@services': resolve(__dirname, 'src/services'),
+      '@tests': resolve(__dirname, 'src/__tests__')
+    }
+  },
+  test: {
+    globals: true,
+    environment: 'jsdom',
+    setupFiles: 'src/setupTests.js'
+  },
+})
