--- src/advanced/__tests__/advanced.test.js
@@ -1,7 +1,0 @@
-import {describe, expect, it} from "vitest";
-
-describe('advanced test', () => {
-  it('테스트코드를 자유롭게 한번 연습해보세요.', () => {
-    expect(true).toBe(true);
-  });
-});
